requirements.txt
Flask
Pillow
together
python-dotenv
flask-cors
